+++ lib/reg.rs
