+++ lib/shift_reg.rs
